Update README.md
Just deleted the general sentence from the first section in the readme.me. I would rather write more concrete what the connector is doing, but I would need more information to do that.

Could you explain me what the connector is doing and what an enduser can do with it?